continuous inference with image saving upon detection
this does not use a pir sensor or activate lora, but it will only
save an image if there is a detection of interest. can be used to
test the functionality but also used to collect images for model
training with potentially less domain shift. this can be used for
power analysis and also allow for easier testing of other options
when it comes to sleep functions and ensuring the pir does not
affect the sd card writing