full tinyscrubcam functionality
now this code successfully will wait for a pir trigger, and
begin inferencing upon the trigger for 10 seconds. if a rhino
is detected, it will write the photo to the sd card and send pin
13 high (which will be attached to the feather, where it will be
waiting for the trigger to send a lora alert) and then make this
pin low and continue waiting for pir trigger after. it is able
to continue this process whether or not it detects a rhino