fix bugs with triggering lora
previously it worked with lighting an led as a placeholder for the
lora component, but serial transmission is required as well other
wise the lora triggers often from false highs. this way other
inputs could be sent as the lora alert as well, if we ever detected
multiple classes. also fixed another bug that made this continuously
trigger from putting the pir sensor as input_pullup