update feather frame counter
the feathers frame counter would reset with each detection. this
causes issues on the server side, where the frame counter sent
in the device payload needs to be higher than the last frame
count logged. It will reject pings that have a frame count
equal to or lower than the last one logged. To resolve this, the
feather will now read and write a variable to flash storage that
keeps the last frame count regardless of if the device restarts.
This way the frame count of a lora packet will always be higher than
the last one received by the gateway. Also increased TX interval
so that multiple packets wouldn't be sent in the 3 minute window.
Also set the first f count value at 20. This realistically should
be 0, but I need this code to work for Saruni's device in kenya
which already has a higher-than-0 frame count, and rather than
generate and add new keys to the gateway server (its a bit of
a hassle) I just want the next ping to exceed the frame count
last logged from this device on the gateway server. Also, an
improvement can be made to this code to introduce wear-leveling,
which will alternate the bin in which the variable is stored
so the flash doesn't wear out as quickly. But for now it works.

Resolves #13